requirements.txt
FastAPI
uvicorn
PyJWT
pydantic
sqlalchemy
psycopg2-binary
passLib
argon2_cffi